+++ tests/TestCase.php
@@ -3,11 +3,8 @@
 namespace MongoDB\Tests;
 
 use InvalidArgumentException;
-use MongoDB\Client;
-use MongoDB\Driver\Manager;
 use MongoDB\Driver\ReadConcern;
 use MongoDB\Driver\ReadPreference;
-use MongoDB\Driver\ServerApi;
 use MongoDB\Driver\WriteConcern;
 use MongoDB\Model\BSONArray;
 use MongoDB\Model\BSONDocument;
@@ -42,16 +39,6 @@ abstract class TestCase extends BaseTestCase
     public static function getUri()
     {
         return getenv('MONGODB_URI') ?: 'mongodb://127.0.0.1:27017';
-    }
-
-    public static function createTestClient(string $uri = null, array $options = [], array $driverOptions = []) : Client
-    {
-        return new Client($uri ?? static::getUri(), $options, static::appendServerApiOption($driverOptions));
-    }
-
-    public static function createTestManager(string $uri = null, array $options = [], array $driverOptions = []) : Manager
-    {
-        return new Manager($uri ?? static::getUri(), $options, static::appendServerApiOption($driverOptions));
     }
 
     /**
@@ -313,15 +300,6 @@ abstract class TestCase extends BaseTestCase
         return array_map(function ($value) {
             return [$value];
         }, $values);
-    }
-
-    private static function appendServerApiOption(array $driverOptions) : array
-    {
-        if (getenv('API_VERSION') && ! isset($driverOptions['serverApi'])) {
-            $driverOptions['serverApi'] = new ServerApi(getenv('API_VERSION'));
-        }
-
-        return $driverOptions;
     }
 
     /**
